+++ src/Driver/QueueAwareInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Efabrica\HermesExtension\Driver;
+
+interface QueueAwareInterface
+{
+    /**
+     * @return array<int, string>
+     */
+    public function getQueue(): array;
+}