+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- HermesWorker can now accept priorities restriction
+
 ## [1.2.1] - 2025-01-10
 ### Fixed
 - max items processing in RedisProxyListDriver